Class_04_Agentic_SDK_Part_01/pyproject.toml
[project]
name = "class-04-agentic-sdk-part-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.1.0",
]
